Show view dialog after creating a new payee
