+++ budgetmanagerpwa/src/containers/PayeeList.jsx
@@ -3,20 +3,57 @@ import { Button, Container, List } from '@mui/material'
 import AddIcon from '@mui/icons-material/Add'
 import PayeeForm from '../components/PayeeForm'
 import PayeeListItem from '../components/PayeeListItem'
-import { useCreatePayeeMutation, useGetPayeesQuery } from '../redux/apiSlice'
+import {
+  useCreatePayeeMutation,
+  useGetPayeesQuery,
+  useDeletePayeeMutation,
+  useUpdatePayeeMutation,
+} from '../redux/apiSlice'
+import PayeeViewDialog from '../components/PayeeViewDialog'
+import DeleteConfirmation from '../components/DeleteConfirmation'
 
 export default function PayeeList() {
   const [createOpen, setCreateOpen] = React.useState(false)
   const [page, setPage] = React.useState(0)
   const query = useGetPayeesQuery(page)
   const [createPayee] = useCreatePayeeMutation()
+  const [updatePayee] = useUpdatePayeeMutation()
+  const [viewOpen, setViewOpen] = React.useState(false)
+  const [viewPayee, setViewPayee] = React.useState(null)
+  const [editOpen, setEditOpen] = React.useState(false)
+  const [deleteOpen, setDeleteOpen] = React.useState(false)
+  const [editData, setEditData] = React.useState(null)
+  const [deletePayee] = useDeletePayeeMutation()
 
   if (query.isLoading) return <p>Loading...</p>
   const list = query.data
 
-  const onCreateSubmit = (data) => {
-    createPayee(data)
-    return null
+  const onEdit = (data) => {
+    setViewOpen(false)
+    setEditData(data)
+    setEditOpen(true)
+  }
+  const onSubmit = async (oldPayee, payee) => {
+    payee.id = oldPayee.id
+    await updatePayee(payee)
+    setEditOpen(false)
+    setViewPayee(payee.id)
+    setViewOpen(true)
+  }
+  const onDeleteSubmit = async () => {
+    await deletePayee(viewOpen)
+    setViewOpen(false)
+  }
+
+  const onCreateSubmit = async (oldData, data) => {
+    const payeeData = await createPayee(data)
+    setViewPayee(payeeData.data.id)
+    setViewOpen(true)
+  }
+
+  const onItemClick = (id)=>{
+    setViewPayee(id)
+    setViewOpen(true)
   }
 
   return (
@@ -27,7 +64,7 @@ export default function PayeeList() {
       {list.count ? (
         <List>
           {list.results.map((item) => (
-            <PayeeListItem item={item} key={item.id} />
+            <PayeeListItem item={item} key={item.id} onClick={onItemClick} />
           ))}
         </List>
       ) : (
@@ -41,6 +78,25 @@ export default function PayeeList() {
         onSubmit={onCreateSubmit}
         open={createOpen}
         title="Add payee"
+      />
+      <PayeeViewDialog
+        open={viewOpen}
+        onClose={() => setViewOpen(false)}
+        payeeId={viewPayee}
+        onEdit={onEdit}
+        onDelete={() => setDeleteOpen(true)}
+      />
+      <PayeeForm
+        payee={editData}
+        open={editOpen}
+        onClose={() => setEditOpen(false)}
+        onSubmit={onSubmit}
+      />
+      <DeleteConfirmation
+        onClose={() => setDeleteOpen(false)}
+        onSubmit={onDeleteSubmit}
+        open={deleteOpen}
+        title="Are you sure you want to delete this payee?"
       />
     </Container>
   )
